requirements.txt
Django==3.2.7
gunicorn==20.1.0
whitenoise==5.3.0
django-crispy-forms
telepot
django-taggit
dj-database-url